backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
supabase>=2.10.0
aiohttp==3.9.1
python-dotenv==1.0.0
httpx>=0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-genai==0.2.0
